LM-Fuser/requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohttp @ file:///croot/aiohttp_1715108769648/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annotated-types==0.7.0
anyio==4.4.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///work/perseverance-python-buildout/croot/attrs_1698845858439/work
av==13.0.0
bitsandbytes==0.43.2
blis==0.7.11
Bottleneck @ file:///croot/bottleneck_1709069899917/work
Brotli @ file:///croot/brotli-split_1714483155106/work
catalogue==2.0.10
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpathlib==0.18.1
comm @ file:///croot/comm_1709322850197/work
confection==0.1.5
cymem==2.0.8
datasets @ file:///croot/datasets_1716911606380/work
debugpy @ file:///work/perseverance-python-buildout/croot/debugpy_1698884710808/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill @ file:///croot/dill_1715094664823/work
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
evaluate==0.4.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
frozenlist @ file:///work/perseverance-python-buildout/croot/frozenlist_1698886842990/work
fsspec @ file:///croot/fsspec_1714461537038/work
gitdb==4.0.11
GitPython==3.1.43
groq==0.9.0
h11==0.14.0
hf_transfer==0.1.8
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna @ file:///croot/idna_1714398848350/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///croot/ipython_1718287989724/work
jedi @ file:///croot/jedi_1721058342488/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jiter==0.5.0
joblib==1.4.2
jupyter_client @ file:///work/perseverance-python-buildout/croot/jupyter_client_1701731788627/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib-inline @ file:///work/perseverance-python-buildout/croot/matplotlib-inline_1698864771271/work
mdurl==0.1.2
metaspace==0.0.1
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict @ file:///croot/multidict_1701096859099/work
multiprocess @ file:///work/perseverance-python-buildout/croot/multiprocess_1698870874259/work
murmurhash==1.0.10
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1720002482208/work
nltk==3.9.1
numexpr @ file:///work/perseverance-python-buildout/croot/numexpr_1698871031164/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp312-cp312-linux_x86_64.whl#sha256=92c518f7e09edd50b5caa5862636c51d6a29391803f3ada62f68aa52f27d8f92
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.12.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow @ file:///croot/pillow_1721059439630/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
preshed==3.0.9
prettytable==3.11.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==3.20.3
psutil @ file:///work/perseverance-python-buildout/croot/psutil_1698863411559/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow @ file:///croot/pyarrow_1721664224170/work/python
pydantic==2.8.2
pydantic_core==2.20.1
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex @ file:///work/perseverance-python-buildout/croot/regex_1698863574928/work
requests @ file:///croot/requests_1721410876868/work
rich==13.7.1
rouge_score==0.1.2
safetensors @ file:///croot/safetensors_1719507957558/work
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
setuptools==69.5.1
shellingham==1.5.4
shtab==1.7.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy @ file:///croot/sympy_1701397643339/work
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers @ file:///croot/tokenizers_1721139552427/work
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado @ file:///croot/tornado_1718740109488/work
tqdm @ file:///croot/tqdm_1716395931952/work
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.44.2
triton==3.0.0
typer==0.12.4
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tyro==0.8.6
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.3
uvloop==0.19.0
wandb==0.17.6
wasabi==1.1.3
watchfiles==0.22.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weasel==0.4.1
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
xxhash @ file:///work/perseverance-python-buildout/croot/python-xxhash_1698900334985/work
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///work/perseverance-python-buildout/croot/zipp_1707367835224/work
